refactor(admin): modularize courses admin config
- Split monolithic admin.py into a package
- Grouped related admin components
- Created dedicated files for each model admin
- Enhances maintainability and readability
- Improves code organization
